Created Base model for User to inherit
