+++ tasks/src/main/java/com/service/tasks/models/Base.java
@@ -1,0 +1,45 @@
+package com.service.tasks.models;
+
+import javax.persistence.*;
+
+import java.io.Serializable;
+import java.time.ZonedDateTime;
+
+@MappedSuperclass
+public abstract class Base implements Comparable<Base>, Serializable {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id")
+    private Long id;
+
+    @Column(name="created_at", nullable = false)
+    private ZonedDateTime createdAt;
+
+    @Column(name="updated_at", nullable = false)
+    private ZonedDateTime updatedAt;
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(long _id) {
+        this.id = _id;
+    }
+
+    public ZonedDateTime getCreatedAt() {
+        return createdAt;
+    }
+
+    public void setCreatedAt(ZonedDateTime createdAt) {
+        this.createdAt = createdAt;
+    }
+
+    public ZonedDateTime getUpdatedAt() {
+        return updatedAt;
+    }
+
+    public void setUpdatedAt(ZonedDateTime updatedAt) {
+        this.updatedAt = updatedAt;
+    }
+
+}
